refactor: 모의서버 명세 수정, 유틸 함수 위치 변경
- 모의 서버 api 명세에 리턴값의 key 부분에 간단한 오타를 수정했습니다.

- 모의 서버에서만 사용되는 유틸 함수를 mocks/ 디렉토리 내부로
  이동했습니다.

  ps. 이거 정말 좋네요. 고의로 에러 만들어서 테스트 해 보기도 좋구요. bb
